Refactored tokentrees into their own files in preparation for tokenstreams. Modified tests to point to the new file now.
