+++ src/test/compile-fail-fulldeps/auxiliary/macro_crate_test.rs
@@ -16,11 +16,12 @@ extern crate syntax;
 extern crate rustc;
 extern crate rustc_plugin;
 
-use syntax::ast::{self, TokenTree, Item, MetaItem, ImplItem, TraitItem, ItemKind};
+use syntax::ast::{self, Item, MetaItem, ImplItem, TraitItem, ItemKind};
 use syntax::codemap::Span;
 use syntax::ext::base::*;
 use syntax::parse::{self, token};
 use syntax::ptr::P;
+use syntax::tokenstream::{TokenTree};
 use rustc_plugin::Registry;
 
 #[macro_export]
